move dynamodbrequests to dynamodb package
