+++ src/main/scala/akka/persistence/dynamodb/snapshot/DynamoDBSnapshotRequests.scala
@@ -3,12 +3,13 @@
   */
 package akka.persistence.dynamodb.snapshot
 
-import java.util.{ Collections, HashMap => JHMap, List => JList, Map => JMap }
+import java.util.{Collections, HashMap => JHMap, List => JList, Map => JMap}
 
-import akka.persistence.{ SelectedSnapshot, SnapshotMetadata, SnapshotSelectionCriteria }
-import akka.persistence.dynamodb.journal.DynamoDBRequests
+import akka.persistence.dynamodb.DynamoDBRequests
+import akka.persistence.{SelectedSnapshot, SnapshotMetadata, SnapshotSelectionCriteria}
 import akka.persistence.serialization.Snapshot
 import com.amazonaws.services.dynamodbv2.model._
+
 import collection.JavaConverters._
 import scala.concurrent.Future
